Remove `SortCursor.h` from `MergeTask.h` transitive include chain
Extract `TextIndexSegment` struct from `TextIndexUtils.h` into a separate
lightweight header `TextIndexSegment.h`. Replace `#include <TextIndexUtils.h>`
with `#include <TextIndexSegment.h>` in `MergeTask.h`, since it only needs
the `TextIndexSegment` struct definition (other classes are forward-declared).

This removes the heavy `SortCursor.h` header (containing a 55-type
`SortQueueVariants` std::variant) from the include chain of `MergeTask.h`
→ `MergeTreeDataMergerMutator.h` → `StorageReplicatedMergeTree.h`, which
is included by ~44 .cpp files. Saves ~1.3s of `SortQueueVariants` parse
time per affected TU, totaling ~196s of CPU time reduction (652m vs 655m).

Specific improvements:
- `StorageSystemReplicas.cpp.o`: 20.9s → 16.6s (-21%)
- `StorageSystemReplicationQueue.cpp.o`: 19.9s → 13.8s (-31%)
- `StorageSystemPartMovesBetweenShards.cpp.o`: 18.3s → 13.5s (-26%)
